requirements-topics.txt
# Web scraping
selenium==4.15.2
beautifulsoup4==4.12.2
webdriver-manager==4.0.1

# HTTP requests and processing
requests==2.31.0
urllib3==2.0.7

# Database
supabase==2.0.3

# Scheduling
schedule==1.2.1

# Other utilities
python-dotenv==1.0.0 